+++ src/AppBundle/Controller/DefaultController.php
@@ -3,20 +3,15 @@
 namespace AppBundle\Controller;
 
 
-use AppBundle\Entity\Contact;
 use AppBundle\Entity\EmailNewsletter;
 use AppBundle\Entity\Taxrefv10;
 use AppBundle\Entity\Observation;
 use AppBundle\Form\EmailNewsletterType;
 use AppBundle\Form\ObservationFrontType;
-use FOS\UserBundle\Event\GetResponseUserEvent;
-use FOS\UserBundle\FOSUserEvents;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
-use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
-use AppBundle\Form\ContactType;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Security;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Method;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
@@ -222,77 +217,7 @@ class DefaultController extends Controller
         ));
     }
 
-    /**
-     *
-     * @Route("/learnmore", name="learn_more")
-     * @return \Symfony\Component\HttpFoundation\Response
-     * @Method({"GET"})
-     *
-     */
-    public function learnMoreAction()
-    {
-        return $this->render('AppBundle:Front:learnMore.html.twig');
-    }
 
-    /**
-     *
-     * @Route("/fonctionnement", name="fonctionnement")
-     * @param Request $request
-     * @return \Symfony\Component\HttpFoundation\Response
-     * @Method({"GET", "POST"})
-     *
-     */
-    public function fonctionnementAction(Request $request)
-    {
-
-        return $this->render('AppBundle:Front:fonctionnement.html.twig');
-    }
-
-
-    /**
-     *
-     * @Route("/legalNotice", name="legal_notice")
-     * @return \Symfony\Component\HttpFoundation\Response
-     * @Method({"GET"})
-     *
-     */
-    public function legalNoticeAction()
-    {
-        return $this->render('AppBundle:Front:legalNotice.html.twig');
-    }
-
-
-
-
-    /**
-     *
-     * @Route("/landing", name="landing")
-     * @return \Symfony\Component\HttpFoundation\Response
-     * @Method({"GET"})
-     *
-     */
-    public function landingAction(Request $request)
-    {
-        $formFactory = $this->get('fos_user.registration.form.factory');
-        $userManager = $this->get('fos_user.user_manager');
-        /** @var $dispatcher EventDispatcherInterface */
-        $dispatcher = $this->get('event_dispatcher');
-
-        $user = $userManager->createUser();
-        $user->setEnabled(true);
-
-        $event = new GetResponseUserEvent($user, $request);
-        $dispatcher->dispatch(FOSUserEvents::REGISTRATION_INITIALIZE, $event);
-
-        if (null !== $event->getResponse()) {
-            return $event->getResponse();
-        }
-
-        $form = $formFactory->createForm();
-        $form->setData($user);
-        return $this->render('AppBundle:Landing:index.html.twig', array (
-            'form' => $form->createView()));
-    }
 
     /**
      * @Route("/viewmyobservation/{page}", name="view_my_observation")
@@ -391,95 +316,4 @@ class DefaultController extends Controller
             throw new NotFoundHttpException("La page demandée n'existe pas");
         }
     }
-
-    /**
-     *
-     * @Route("/contact", name="modal_contact")
-     * @param Request $request
-     * @return \Symfony\Component\HttpFoundation\Response
-     * @Method({"GET", "POST"})
-     *
-     */
-    public function modalContactAction(Request $request)
-    {
-        $contact = new Contact();
-        $form = $this->get('form.factory')->create(ContactType::class, $contact);
-        $form->handleRequest($request);
-        if ($form->isSubmitted() && $form->isValid()) {
-            // Affichage d'un message flash
-            $request->getSession()->getFlashBag()->add('success', 'Votre message à bien été envoyé !');
-            // Sauvegarder en Base de données
-            $em = $this->getDoctrine()->getManager();
-            $em->persist($contact);
-            $em->flush();
-            // Envoyer le mail
-            $this->container->get('app.sendEmail')->sendEmailContact($contact);
-            // Vider l'objet contact
-            $contact = null;
-            // Retour à la page d'accueil
-            $em = $this->getDoctrine()->getManager();
-            $listLastObservations = $em->getRepository('AppBundle:Observation')->findLastObservations(3);
-
-            return $this->render('AppBundle:Front:index.html.twig', array(
-                'listLastObservations' => $listLastObservations
-            ));
-        }
-        return $this->render('AppBundle:Front:modalContact.html.twig', array(
-            'form' => $form->createView(),
-        ));
-    }
-
-
-    /**
-     *
-     * @return \Symfony\Component\HttpFoundation\Response
-     * @param Request $request
-     * @Route("/sidebarnewsletter", name="side_bar_newsletter")
-     *
-     */
-    public function sideBarNewsletterAction(Request $request)
-    {
-        $emailNewsletter = new EmailNewsletter();
-        // On crée le formulaire
-        $form = $this->createForm(EmailNewsletterType::class, $emailNewsletter);
-
-        $form->handleRequest($request);
-        if ($form->isSubmitted() && $form->isValid()) {
-            // Affichage d'un message flash
-
-            $request->getSession()->getFlashBag()->add('success', 'Vous êtes désormais inscrit à notre Newsletter');
-
-            // Sauvegarder en Base de données
-            $em = $this->getDoctrine()->getManager();
-            $em->persist($emailNewsletter);
-            $em->flush();
-
-            // Retour à la page d'accueil
-            $em = $this->getDoctrine()->getManager();
-            $listLastObservations = $em->getRepository('AppBundle:Observation')->findLastObservations(3);
-
-            return $this->render('AppBundle:Front:index.html.twig', array(
-                'listLastObservations' => $listLastObservations,
-                'form' => $form->createView()
-            ));
-        }
-    }
-
-    /**
-     *
-     * @return \Symfony\Component\HttpFoundation\Response
-     * @Route("/sidebarlast", name="side_bar_last")
-     * @Method({"GET"})
-     *
-     */
-    public function sideBarLastAction()
-    {
-        $em = $this->getDoctrine()->getManager();
-        $listLastObservations = $em->getRepository('AppBundle:Observation')->findLastObservations(10);
-        return $this->render('AppBundle:Front:sideBarLast.html.twig', array(
-            'listLastObservations' => $listLastObservations
-        ));
-    }
-
-
 }
