+++ src/AppBundle/Resources/views/layout.html.twig
@@ -29,7 +29,7 @@
 
     {% block body %}{% endblock %}
 
-    {{ render(controller('AppBundle:Default:modalContact')) }}
+    {{ render(controller('AppBundle:Nav:modalContact')) }}
     {{ include("AppBundle:Front:footer.html.twig") }}
